requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
sqlalchemy 
pydantic 
bcrypt
python-jose[cryptography]
python-multipart
python-jose
